refactor: extract useFaviconState composable from App.vue
faviconState computed + notification badge + useDynamicFavicon call.
